+++ shuup/admin/modules/orders/__init__.py
@@ -45,7 +45,7 @@ class OrderModule(AdminModule):
             admin_url(
                 r"^shipments/$",
                 "shuup.admin.modules.orders.views.ShipmentListView",
-                name="order.shipments-list"
+                name="order.shipments.list"
             ),
             admin_url(
                 r"^orders/(?P<pk>\d+)/create-payment/$",
@@ -132,9 +132,8 @@ class OrderModule(AdminModule):
             ),
             MenuEntry(
                 text=_("Shipments"),
-                url="shuup_admin:order.shipments-list",
+                url="shuup_admin:order.shipments.list",
                 category=ORDERS_MENU_CATEGORY,
-                aliases=[_("Shipments list")], 
                 ordering=2
             )
         ]
@@ -219,9 +218,3 @@ class OrderStatusModule(AdminModule):
 
     def get_model_url(self, object, kind, shop=None):
         return derive_model_url(OrderStatus, "shuup_admin:order_status", object, kind)
-
-
-class ShipmentModule(AdminModule):
-    name = _("Shipments")
-    breadcrumbs_menu_entry = MenuEntry(name, url="shuup_admin:order.shipments-list")
-