+++ shuup/admin/modules/orders/views/__init__.py
@@ -9,14 +9,15 @@
 from .addresses import OrderAddressEditView
 from .detail import OrderDetailView, OrderSetStatusView
 from .edit import OrderEditView, UpdateAdminCommentView
-from .list import OrderListView, ShipmentListView
+from .list import OrderListView
 from .log import NewLogEntryView
 from .payment import (
     OrderCreatePaymentView, OrderDeletePaymentView, OrderSetPaidView
 )
 from .refund import OrderCreateFullRefundView, OrderCreateRefundView
 from .shipment import (
-    OrderCreateShipmentView, ShipmentDeleteView, ShipmentSetSentView
+    OrderCreateShipmentView, ShipmentDeleteView, ShipmentListView,
+    ShipmentSetSentView
 )
 from .status import OrderStatusEditView, OrderStatusListView